Revert "Witness functions!!! (#6064)"
This reverts commit 23278a5f65cecfe2f9bc533c8d461fd5bdeca93c.
