+++ drake/systems/framework/BUILD
@@ -227,9 +227,7 @@ drake_cc_library(
 drake_cc_library(
     name = "system",
     srcs = ["system.cc"],
-    hdrs = ["system.h",
-            "discrete_event.h",
-            "witness_function.h"],
+    hdrs = ["system.h"],
     deps = [
         ":cache",
         ":context",
